+++ config.inc.php
@@ -10,8 +10,8 @@ $REX['ADDON']['perm']['rexseo42'] = 'rexseo42[]';
 
 // permissions
 $REX['PERM'][] = 'rexseo42[]';
-$REX['EXTPERM'][] = 'rexseo42[seopage]';
-$REX['EXTPERM'][] = 'rexseo42[seopage_extended]';
+$REX['EXTPERM'][] = 'rexseo42[seo_default]';
+$REX['EXTPERM'][] = 'rexseo42[seo_extended]';
 
 // includes
 require($REX['INCLUDE_PATH'] . '/addons/rexseo42/classes/class.rexseo42.inc.php');
@@ -38,7 +38,7 @@ if ($REX['REDAXO']) {
 	);
 
 	// check for user permissions (admins will have all)
-	if (isset($REX['USER']) && ($REX['USER']->isAdmin() || $REX['USER']->hasPerm('rexseo42[seopage]') || $REX['USER']->hasPerm('rexseo42[seopage_extended]') || $REX['USER']->hasPerm('editContentOnly[]'))) {
+	if (isset($REX['USER']) && ($REX['USER']->isAdmin() || $REX['USER']->hasPerm('rexseo42[seo_default]') || $REX['USER']->hasPerm('rexseo42[seo_extended]') || $REX['USER']->hasPerm('editContentOnly[]'))) {
 		// react on one_page_mode option
 		if (!$REX['ADDON']['rexseo42']['settings']['one_page_mode'] || ($REX['ADDON']['rexseo42']['settings']['one_page_mode'] && $REX['ARTICLE_ID'] == $REX['START_ARTICLE_ID'])) {
 			// add new menu item